fix: properly terminate stuck runs
