+++ bento_wes/backends/__init__.py
@@ -1,10 +1,8 @@
 from . import backend_types
-from ._finish_run import finish_run
 from ._wes_backend import WESBackend
 
 
 __all__ = [
     "backend_types",
-    "finish_run",
     "WESBackend",
 ]
